app/requirements.txt
flask
waitress
mysql-connector-python
requests
imagekitio
